Split lua file test code into header
